Fix replyTo JSON deserialization error for negative packet IDs
Change ReplyTo type from *uint32 to *int64 to handle signed JSON numbers
since packet IDs can appear as negative when serialized.

+++ meshtastic-go/internal/api/handlers/messages.go
@@ -75,7 +75,7 @@ type SendMessageRequest struct {
 	To      interface{} `json:"to" binding:"required"` // Can be string "!hex" or number
 	Channel int         `json:"channel"`
 	Text    string      `json:"text" binding:"required"`
-	ReplyTo *uint32     `json:"replyTo,omitempty"` // Packet ID of message being replied to
+	ReplyTo *int64      `json:"replyTo,omitempty"` // Packet ID of message being replied to (int64 to handle signed JSON numbers)
 }
 
 // SendMessage sends a message
